requirements.txt
asgiref==3.3.0
attrs==20.3.0
Automat==20.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
cssselect==1.1.0
Cython==0.29.14
defusedxml==0.7.0rc1
Django==3.1.3
django-cors-headers==3.5.0
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
djoser==2.1.0
drf-yasg==1.20.0
drfauto==1.3
gensim==3.8.3
hyperlink==20.0.1
idna==2.10
importlib-metadata==1.7.0
incremental==17.5.0
inflection==0.5.1
itemadapter==0.2.0
itemloaders==1.0.4
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==1.0.1
loguru==0.5.3
lxml==4.6.2
MarkupSafe==1.1.1
nltk==3.5
numpy==1.20.1
oauthlib==3.1.0
packaging==20.4
parsel==1.6.0
Protego==0.1.16
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
PyJWT==1.7.1
pymystem3==0.1.10
pyOpenSSL==20.0.0
pyparsing==2.4.7
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2020.4
queuelib==1.5.0
regex==2021.3.17
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scikit-learn==0.24.1
scipy==1.6.2
Scrapy==2.4.1
service-identity==18.1.0
six==1.15.0
sklearn==0.0
smart-open==4.2.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.0.1
sqlparse==0.4.1
threadpoolctl==2.1.0
tqdm==4.59.0
Twisted==20.3.0
uritemplate==3.0.1
urllib3==1.25.11
w3lib==1.22.0
win32-setctime==1.0.3
zipp==3.4.0
zope.interface==5.2.0
